Move payload unit tests to test_module_api.cpp
